add identity, a library using Microsoft Identity to handle user auth
tempted to change it at a later point to be a separate IdP that's independent of the api and app that can be used to handle auth.  The only problem of note is how to handle registration in that case, since that is app specific.

Chances are we'd let the external IdP handle its own local registration and when user logs in the app or api that allows us to handle registration from that point on, so we're only 'creating' an account within our app but the authentication is still handled by separate IdP